refactor(scala): split package.scala into separate traits
